Consolidating configs
