--- web/next.config.js
@@ -1,17 +1,0 @@
-/** @type {import('next').NextConfig} */
-
-const isDev = process.env.NODE_ENV === "development";
-
-const nextConfig = {
-  ...(isDev && {
-    webpackDevMiddleware: (config) => {
-      config.watchOptions = {
-        poll: 1000,
-        aggregateTimeout: 300,
-      };
-      return config;
-    },
-  }),
-};
-
-module.exports = nextConfig;
